requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.13.1
requests==2.28.1
pytest==7.1.2
